Replace run_rdt into main file
