+++ todo.md
@@ -5,8 +5,10 @@
 ### rna_dna_tools
 
 - [ ] Replace nucleic acids type check into functions
-- [ ] Add count_gc() and translate_sequence()
+- [ ] Add count_gc()
+- [ ] Add translate_sequence()
 - [ ] Add dictionary to call functions
+- [ ] Add docstrings of functions
   
 ### readme_md
